% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{A helper function for simViz}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Your plots}

\item{plotlist}{A list objects of plots}

\item{file}{The variable you want on the y-axis}

\item{cols}{The number of columns}

\item{layout}{More specific layout stuff}
}
\value{
A panel with several visualizations.
}
\description{
A helper function for simViz
}
\examples{
simViz(mtcars, mpg, cyl, 2)
}
