% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMod.R
\name{simMod}
\alias{simMod}
\title{Create a table of model coefficients}
\usage{
simMod(dat, dv, ..., distractors = 3, answer = FALSE)
}
\arguments{
\item{dat}{Your data}

\item{dv}{The bare variable you want as the dv in your model}

\item{...}{The bare variables you want as the predictors in your model}

\item{distractors}{The number of distractor coefficients you want included. Defaults to 3.}

\item{answer}{Logical. TRUE if you want to play the game, FALSE otherwise.}
}
\value{
A data.frame with several sets of coefficients.
}
\description{
Create a table of model coefficients
}
\examples{
simMod(mtcars, mpg, hp, cyl)
}
